Detailed instruction, how to setup xDebug
Using NFQ Kickstarter for infrastructure examples,
because those instructions were written for students to get xDebug easier.

Added as much images as possible, so it would be easier to follow for not PHP developers.

Included more output in xDebug scripts, so end use will know, that right command was executed.